refactor: remove useless interface file
